# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AccountPool": {
        "additionalProperties": false,
        "description": "一个 Environment 中按角色组织的独占账号调度池。",
        "properties": {
          "cooldownSeconds": {
            "maximum": 86400.0,
            "minimum": 0.0,
            "title": "Cooldownseconds",
            "type": "integer"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "defaultTtlSeconds": {
            "maximum": 7200.0,
            "minimum": 300.0,
            "title": "Defaultttlseconds",
            "type": "integer"
          },
          "environmentId": {
            "format": "uuid",
            "title": "Environmentid",
            "type": "string"
          },
          "exclusive": {
            "title": "Exclusive",
            "type": "boolean"
          },
          "healthFailureThreshold": {
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Healthfailurethreshold",
            "type": "integer"
          },
          "healthRetryCooldownSeconds": {
            "maximum": 86400.0,
            "minimum": 0.0,
            "title": "Healthretrycooldownseconds",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "poolKey": {
            "title": "Poolkey",
            "type": "string"
          },
          "projectId": {
            "format": "uuid",
            "title": "Projectid",
            "type": "string"
          },
          "revision": {
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          },
          "roleId": {
            "format": "uuid",
            "title": "Roleid",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AccountPoolStatus"
          },
          "tenantId": {
            "format": "uuid",
            "title": "Tenantid",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenantId",
          "projectId",
          "environmentId",
          "roleId",
          "poolKey",
          "name",
          "exclusive",
          "defaultTtlSeconds",
          "cooldownSeconds",
          "healthFailureThreshold",
          "healthRetryCooldownSeconds",
          "status",
          "revision",
          "createdAt",
          "updatedAt"
        ],
        "title": "AccountPool",
        "type": "object"
      },
...
      "AccountPoolStatus": {
        "description": "账号池调度状态。",
        "enum": [
          "ACTIVE",
          "DISABLED"
        ],
        "title": "AccountPoolStatus",
        "type": "string"
      },
...
      "ProblemDetails": {
        "additionalProperties": false,
        "description": "Atlas 扩展的 RFC Problem Details。",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errorCode": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "instance": {
            "title": "Instance",
            "type": "string"
          },
          "requestId": {
            "title": "Requestid",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "violations": {
            "items": {
              "$ref": "#/components/schemas/FieldViolation"
            },
            "title": "Violations",
            "type": "array"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "errorCode",
          "requestId"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Atlas TestOps Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/account-pools/{poolId}": {
      "get": {
        "operationId": "get_account_pool_v1_account_pools__poolId__get",
        "parameters": [
          {
            "in": "path",
            "name": "poolId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Poolid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Atlas-Tenant-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Tenant-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Atlas-Actor-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Actor-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountPool"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "请求语义或 Revision 无效"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "缺少有效身份"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "当前 PlatformRole 无权执行"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "资源不存在或不可见"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "唯一键、幂等或状态转换冲突"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Revision 前置条件失败"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "请求不符合接口契约"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "服务内部错误"
          }
        },
        "summary": "读取 AccountPool",
        "tags": [
          "identity"
        ]
      },
      "patch": {
        "operationId": "update_account_pool_v1_account_pools__poolId__patch",
        "parameters": [
          {
            "in": "path",
            "name": "poolId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Poolid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "maxLength": 64,
              "title": "If-Match",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Atlas-Tenant-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Tenant-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Atlas-Actor-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Actor-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountPool"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountPool"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "请求语义或 Revision 无效"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "缺少有效身份"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "当前 PlatformRole 无权执行"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "资源不存在或不可见"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "唯一键、幂等或状态转换冲突"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Revision 前置条件失败"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "请求不符合接口契约"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "服务内部错误"
          }
        },
        "summary": "更新 AccountPool",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/account-pools/{poolId}/accounts": {
      "get": {
        "description": "只返回脱敏账号提示和计算可用性，不返回 CredentialRef。",
        "operationId": "list_test_accounts_v1_account_pools__poolId__accounts_get",
        "parameters": [
          {
            "in": "path",
            "name": "poolId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Poolid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Atlas-Tenant-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Tenant-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Atlas-Actor-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Actor-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestAccountPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "请求语义或 Revision 无效"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "缺少有效身份"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "当前 PlatformRole 无权执行"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "资源不存在或不可见"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "唯一键、幂等或状态转换冲突"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Revision 前置条件失败"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "请求不符合接口契约"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "服务内部错误"
          }
        },
        "summary": "列出 TestAccount",
        "tags": [
          "identity"
        ]
      },
      "post": {
        "description": "原子导入账号元数据、独占 Slot 与不可兑换的 SecretRef。",
        "operationId": "create_test_account_v1_account_pools__poolId__accounts_post",
        "parameters": [
          {
            "in": "path",
            "name": "poolId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Poolid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Atlas-Tenant-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Tenant-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Atlas-Actor-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Actor-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTestAccount"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestAccount"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "请求语义或 Revision 无效"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "缺少有效身份"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "当前 PlatformRole 无权执行"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "资源不存在或不可见"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "唯一键、幂等或状态转换冲突"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Revision 前置条件失败"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "请求不符合接口契约"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "服务内部错误"
          }
        },
        "summary": "导入 TestAccount",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/account-pools/{poolId}/capacity": {
      "get": {
        "description": "实时计算可用、租用、冷却、隔离与未验证容量。",
        "operationId": "get_account_pool_capacity_v1_account_pools__poolId__capacity_get",
        "parameters": [
          {
            "in": "path",
            "name": "poolId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Poolid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Atlas-Tenant-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Tenant-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Atlas-Actor-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Actor-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountPoolCapacity"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "请求语义或 Revision 无效"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "缺少有效身份"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "当前 PlatformRole 无权执行"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "资源不存在或不可见"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "唯一键、幂等或状态转换冲突"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Revision 前置条件失败"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "请求不符合接口契约"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "服务内部错误"
          }
        },
        "summary": "读取 AccountPool 容量",
        "tags": [
          "identity"
        ]
      }
    },
...
  }
}